+++ src/config.rs
@@ -1,7 +1,11 @@
 use std::path::PathBuf;
 
+use super::format::Format;
+
 #[derive(Debug)]
 pub struct Config {
+    pub input_format: Format,
+    pub output_format: Format,
     pub timestamp: std::time::SystemTime,
     pub uid: u32,
     pub gid: u32,
@@ -43,6 +47,8 @@ impl Config {
 impl Default for Config {
     fn default() -> Self {
         Config {
+            input_format: Format::Json,
+            output_format: Format::Json,
             timestamp: std::time::SystemTime::now(),
             uid: 501,
             gid: 501,
